move helper functions over
